requirements.txt
streamlit==1.11.1
Pillow==9.2.0
pandas==1.4.3
matplotlib==3.5.2
streamlit-player==0.1.5
PyQRCode==1.2.1
pypng==0.20220715.0